Allow subclassing InvitationsAdapter
Previously, the code would check to see if ACCOUNT_ADAPTER equalled
"invitations.models.InvitationsAdapter". But this would cause it to
reject subclasses of InvitationsAdapter. Instead, it is enough to check
to see if ACCOUNT_ADAPTER setting exists, as this setting is meant for
django-allauth.

Modify the tests to check for this.